requirements.txt
Django==1.11.7
enum34==1.1.6
djangorestframework==3.7.3
